etlpipeline/requirements.txt
Flask == 3.0.0
pandas == 2.2.1
requests == 2.31.0
sqlalchemy==2.0.32
psycopg2-binary==2.9.9
google-cloud-bigquery==3.25.0
click==8.1.7
